requirements.txt
-i https://pypi.org/simple
--trusted-host pypi.org
adal==1.2.1
afkak==3.0.0
aiohttp==3.5.4
aiojobs==0.2.2
aiokafka==0.5.0
aioslacker==0.0.11
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
automat==0.7.0
boto3==1.9.92
botocore==1.12.92
cachetools==3.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.5
cython==0.29.5
docutils==0.14
google-auth==1.6.2
hyperlink==18.0.0
idna-ssl==1.1.0; python_version < '3.7'
idna==2.8
incremental==17.5.0
jmespath==0.9.3
json-logging==0.0.11
kafka-python==1.4.3
kubernetes-py==1.10.7.1
kubernetes==8.0.1
multidict==4.5.2
numpy==1.16.1
oauthlib==3.0.1
pendulum==2.0.4
pyasn1-modules==0.2.4
pyasn1==0.4.5
pycparser==2.19
pyhamcrest==1.9.0
pyjwt==1.7.1
python-dateutil==2.8.0; python_version >= '2.7'
pytzdata==2018.9
pyyaml==3.13
requests-oauthlib==1.2.0
requests==2.21.0
rsa==4.0
s3transfer==0.2.0
six==1.12.0
slackclient==1.3.0
slacker==0.9.42
twisted==18.9.0
typing-extensions==3.7.2; python_version < '3.7'
urllib3==1.24.1; python_version >= '3.4'
uuid==1.30
vibora==0.0.6
vistir==0.3.0
websocket-client==0.54.0
xlsxwriter==1.1.4
yarl==1.3.0
zope.interface==4.6.0
